requirements.txt
Flask==1.1.2
line-bot-sdk==1.19.0
pymysql==1.0.2